requirements.txt
requests==2.31.0
flask==2.3.3
python-dotenv==1.0.0
pyjwt==2.8.0
cryptography==41.0.4
pytest==7.4.2
